requirements.txt
mlflow>=2.8.0
pymongo>=4.5.0
pyngrok>=6.0.0
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.28.0
pydantic>=1.10.0
python-dotenv>=1.0.0
